Refactor: bloom/ws
